requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap4==2024.1
Django==5.1
django-crispy-forms==2.3
django-mathfilters==1.0.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.2
six==1.16.0
sqlparse==0.5.1
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.1
